// frontend/src/shared/lib/api/model/conditionGroupReadoutResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Chem-Vault
 * OpenAPI spec version: 0.1.0
 */
import type { ConditionGroupReadoutResponseUnit } from './conditionGroupReadoutResponseUnit';

export interface ConditionGroupReadoutResponse {
  readout_definition_id: string;
  name: string;
  value: number;
  unit?: ConditionGroupReadoutResponseUnit;
  aggregation: string;
  count: number;
}