// frontend/src/shared/lib/api/model/interceptValueResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 0.1.0
 */
import type { InterceptSpecResponse } from './interceptSpecResponse';
import type { InterceptValueResponseConfidenceIntervalLow } from './interceptValueResponseConfidenceIntervalLow';
import type { InterceptValueResponseConfidenceIntervalHigh } from './interceptValueResponseConfidenceIntervalHigh';

export interface InterceptValueResponse {
  spec: InterceptSpecResponse;
  value: number;
  confidence_interval_low?: InterceptValueResponseConfidenceIntervalLow;
  confidence_interval_high?: InterceptValueResponseConfidenceIntervalHigh;
  at_bound?: boolean;
}